+++ src/RegistCrop/components/TypeSelect.jsx
