+++ src/RegistCrop/css/TypeSelect.css
