+++ src/RegistCrop/img/left-arrow.png
